First working version of the test
